refactor(util): SeatsFormat util 삭제
