+++ app/src/main/java/woowacourse/movie/presentation/choiceSeat/ChoiceSeatActivity.kt
@@ -21,8 +21,8 @@ import woowacourse.movie.domain.model.tools.seat.SeatRow
 import woowacourse.movie.domain.model.tools.seat.Seats
 import woowacourse.movie.domain.model.tools.seat.Theater
 import woowacourse.movie.presentation.complete.CompleteActivity
+import woowacourse.movie.presentation.mappers.toPresentation
 import woowacourse.movie.presentation.model.ReservationModel
-import woowacourse.movie.presentation.model.toPresentation
 
 class ChoiceSeatActivity : AppCompatActivity() {
