+++ app/src/main/java/woowacourse/movie/presentation/mappers/Mappers.kt
@@ -1,10 +1,12 @@
-package woowacourse.movie.presentation.model
+package woowacourse.movie.presentation.mappers
 
 import woowacourse.movie.R
 import woowacourse.movie.data.MovieDrawableData
 import woowacourse.movie.domain.model.tools.Movie
 import woowacourse.movie.domain.model.tools.Ticket
-import woowacourse.movie.presentation.util.formatLocationToString
+import woowacourse.movie.domain.model.tools.seat.Seat
+import woowacourse.movie.presentation.model.MovieModel
+import woowacourse.movie.presentation.model.TicketModel
 
 fun MovieModel.toDomainModel() = Movie(
     id = id,
@@ -33,5 +35,7 @@ fun Ticket.toPresentation() = TicketModel(
     bookedDateTime = bookedDateTime,
     count = count,
     paymentMoney = getPaymentMoney().value,
-    seats = seats.value.map { it.location }.formatLocationToString(),
+    seats = seats.value.map { it.formatLocation() },
 )
+
+private fun Seat.formatLocation() = "${location.row}${(location.number + 1)}"
