+++ app/src/main/java/woowacourse/movie/presentation/model/MovieModel.kt
@@ -3,6 +3,8 @@ package woowacourse.movie.presentation.model
 import androidx.annotation.DrawableRes
 import woowacourse.movie.domain.model.tools.TicketCount
 import woowacourse.movie.domain.model.tools.seat.Seats
+import woowacourse.movie.presentation.mappers.toDomainModel
+import woowacourse.movie.presentation.mappers.toPresentation
 import java.time.LocalDate
 
 data class MovieModel(
